move utils functions to a helpers directory, change the format command to include unstaged files after format
