+++ src/utils/helpers/escape-regex.ts
