+++ src/utils/helpers/get-chrome-last-error.ts
